requirements.txt
playwright==1.44.0
anthropic==0.28.0
httpx==0.27.0
python-dotenv==1.0.1
